chore: fix formatting issues
