+++ src/core/lock/interval_lock.ts
@@ -5,11 +5,10 @@ import {type K8Factory} from '../kube/k8_factory.js';
 import {LockHolder} from './lock_holder.js';
 import {sleep} from '../helpers.js';
 import {Duration} from '../time/duration.js';
-import {type Lock} from './lock.js';
+import {type Lock, type LockRenewalService} from './lock.js';
 import {StatusCodes} from 'http-status-codes';
 import {type NamespaceName} from '../kube/resources/namespace/namespace_name.js';
 import {type Lease} from '../kube/resources/lease/lease.js';
-import {type LockRenewalService} from './lock_renewal_service.js';
 import {LockAcquisitionError} from './lock_acquisition_error.js';
 import {LockRelinquishmentError} from './lock_relinquishment_error.js';
 import {InjectTokens} from '../dependency_injection/inject_tokens.js';
